Update .nvmrc
Testing what version stable brings - if v5.3.0 I expect a crash, as netlify doesn't have this version.  If it works with an error (specific one)  it's using the default 0.12.x that netlify uses as default and will throw an error when the harmonize flag isn't used! If it works fine and it's using v.5.1.0 everything is as I wanted.